Initial program to parse md files if they're relevant for today
tnotes is small helper to find `*.md` files within a
directory and try to parse them for rules.

When a rule got found it will interpret the current date as:
- ON_WEEKDAY monday-friday
- ON_WEEKEND saturday-sunday
- ON YYYY-MM-DD
- FROM YYYY-MM-DD
- UNTIL YYYY-MM-DD
and display the file as a link to easily open them.

The usage of the program is:

`tnotes /path/to/file_dir`

e.g.:

`tnotes /tmp/example`